data_management/pyproject.toml
[project]
name = "data-management"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0.1",
    "django-admin-sortable2>=2.2.3",
]
